requirements.txt
notion-client==2.2.1
jinja2==3.1.4
weasyprint==60.2
pydyf==0.10.0
python-dotenv==1.0.1
fastapi==0.104.1
uvicorn==0.24.0